lambda_functions/importVideo/requirements.txt
# Copyright 2023

# Licensed under the Apache License, Version 2.0 (the "License");
# you may not use this file except in compliance with the License.
# You may obtain a copy of the License at

# http://www.apache.org/licenses/LICENSE-2.0

# Unless required by applicable law or agreed to in writing, software
# distributed under the License is distributed on an "AS IS" BASIS,
# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
# See the License for the specific language governing permissions and
# limitations under the License.

boto3==1.34.148
botocore==1.34.148
ffmpeg-python==0.2.0
PyMySQL==1.1.1
s3transfer==0.10.2
# FFmpeg & FFprobe added as a layer
